MarkovTable
For the final assignment this semester, you will use your Markov table from the previous step in this assignment to write a full Markov text generator. To generate Markov text, you must:

 

Pull an arbitrary key out of the Markov table
Pull a value at random from the ArrayList of values associated with that key
Print the value
Search for a new key, consisting of every word from the previous key except the first key, and with the value added to the end of the key.
Print a value at random associated with this new key
Repeat until a sufficient number of Markov words has been generated.
Your main function should ask for a number of words to generate, and then use the Markov table from the previous exercise -- using the same source text as in that exercise -- to generate that number of words. 

 

This assignment is due at the last possible second the Canvas shell is open -- Friday at 11:59 PM.