Building a MarkovTable with variable key length
As in the previous assignment, for this one you should create a class called MarkovTable with a method called readFile that takes a String as a parameter (indicating the name of a file to be opened) and returns a boolean (which indicates whether or not the attempt to read the file was successful).

This class should have a private member variable of type HashMap<ArrayList<String>, ArrayList<String>>. 

You'll notice that the map in this version of the assignment associates an ArrayList to an ArrayList. This is because instead of associating a word with the words that can potentially follow it, you are instead associating a group of one or more words with the words that can potentially follow them. Here is an example where the key length is three:

It was the best of times, it was the worst of times, it was the age of wisdom, it was the age of foolishness, it was the epoch of belief, it was the epoch of incredulity, it was the season of Light, it was the season of Darkness, it was the spring of hope, it was the winter of despair, we had everything before us, we had nothing before us, we were all going direct to Heaven, we were all going direct the other way—in short, the period was so far like the present period, that some of its noisiest authorities insisted on its being received, for good or for evil, in the superlative degree of comparison only

Consider the string of words “it was the” (which I’ve bolded). Below is a list of all the words in this paragraph that appear immediately after “it was the”.

[ best, worst, age, age, epoch, epoch, season, season, spring, winter ]

As stated above your key length is 3: you are looking for each individual word that can follow "it was the".

If the key length is 1, the table built should be identical to the table built by your program in the previous assignment in this module. If the key length is two, the value associated with that key is a container containing all possible individual words that can follow each side-by-side pair of words in the sample file. If the key length is three, the value is each individual word that can follow each set of three side-by-side words in the sample file, and so forth.

As in the previous assignment, you should treat the first word in the file as if it followed the last word in the file. For example, if this was your sample text, and if your key length were 2:

one fish two fish red fish blue fish

The key/value pairs in your HashMap would be:

['one fish'] : ['two']
['fish two'] : ['fish']
['two fish'] : ['red']
['fish red'] : ['fish']
['red fish'] : ['blue']
['fish blue'] : ['fish']
['blue fish'] : ['one']
['fish one'] : ['fish']

 

(note: this sample text is too short to be good for a markov generator using a key length of 2 -- the important thing to notice about it is the last two entries in the table)

 

As in the previous example, test your class with this sample file.